refactor(askai): remove Ask AI transport abstraction
