+++ examples/demo-react/src/App.tsx
@@ -7,8 +7,6 @@ import './App.css';
 import '@docsearch/css/dist/style.css';
 import '@docsearch/css/dist/sidepanel.css';
 
-import AgentStudio from './examples/agent-studio';
-import AgentStudioSidepanel from './examples/agent-studio-sidepanel';
 import Basic from './examples/basic';
 import BasicAskAI from './examples/basic-askai';
 import Composable from './examples/composable';
@@ -97,20 +95,6 @@ function App(): JSX.Element {
             <p className="section-description">sidepanel hybrid</p>
             <div className="search-wrapper">
               <BasicHybrid />
-            </div>
-          </section>
-
-          <section className="demo-section">
-            <p className="section-description">Agent Studio</p>
-            <div className="search-wrapper column">
-              <AgentStudio />
-            </div>
-          </section>
-
-          <section className="demo-section">
-            <p className="section-description">Agent Studio sidepanel</p>
-            <div className="search-wrapper column">
-              <AgentStudioSidepanel />
             </div>
           </section>
         </main>
