+++ examples/demo-react/src/examples/basic-askai.tsx
@@ -9,14 +9,28 @@ export default function BasicAskAI(): JSX.Element {
       appId="PMZUYBQDAK"
       apiKey="24b09689d5b4223813d9b8e48563c8f6"
       askAi={{
-        assistantId: 'askAIDemo',
-        searchParameters: {
-          facetFilters: ['language:en'],
-        },
+        assistantId: 'ccdec697-e3fe-465b-a1c3-657e7bf18aef',
         suggestedQuestions: true,
       }}
       insights={true}
       translations={{ button: { buttonText: 'Search with Ask AI' } }}
+      tools={{
+        printConsoleMessage: {
+          render({ message: { output } }) {
+            if (!output) return '';
+
+            return output as string;
+          },
+          async onToolCall({ input, addToolOutput }) {
+            // eslint-disable-next-line no-console
+            console.log((input as any).message);
+
+            await addToolOutput({
+              output: 'Check your console for a nice message :)',
+            });
+          },
+        },
+      }}
     />
   );
 }
