+++ examples/demo-react/src/examples/sidepanel.tsx
@@ -11,8 +11,25 @@ export default function SidepanelExample(): JSX.Element {
         indexName="docsearch"
         appId="PMZUYBQDAK"
         apiKey="24b09689d5b4223813d9b8e48563c8f6"
-        assistantId="askAIDemo"
+        assistantId="ccdec697-e3fe-465b-a1c3-657e7bf18aef"
         variant="floating"
+        tools={{
+          printConsoleMessage: {
+            render({ message: { output } }) {
+              if (!output) return '';
+
+              return output as string;
+            },
+            async onToolCall({ input, addToolOutput }) {
+              // eslint-disable-next-line no-console
+              console.log((input as any).message);
+
+              await addToolOutput({
+                output: 'Check your console for a nice message :)',
+              });
+            },
+          },
+        }}
       />
     </DocSearch>
   );
