+++ packages/docsearch-react/src/constants.ts
@@ -1,4 +1,2 @@
 export const MAX_QUERY_SIZE = 512;
-export const ASK_AI_API_URL = 'https://askai.algolia.com/chat';
-export const BETA_ASK_AI_API_URL = 'https://beta-chat-askai.algolia.com';
 export const SUGGESTED_QUETIONS_INDEX_NAME = 'algolia_ask_ai_suggested_questions';
