+++ packages/docsearch-react/src/useAskAi.ts
@@ -4,21 +4,14 @@ import type { ChatOnToolCallCallback } from 'ai';
 import { DefaultChatTransport, lastAssistantMessageIsCompleteWithToolCalls } from 'ai';
 import { useCallback, useEffect, useMemo, useRef } from 'react';
 
-import {
-  agentStudioBaseUrl,
-  getAgentStudioErrorMessage,
-  getValidToken,
-  postAgentStudioFeedback,
-  postFeedback,
-} from './askai';
+import { agentStudioBaseUrl, getAgentStudioErrorMessage, postAgentStudioFeedback } from './askai';
 import type { Exchange } from './AskAiScreen';
-import { ASK_AI_API_URL, BETA_ASK_AI_API_URL } from './constants';
 import type { StoredSearchPlugin } from './stored-searches';
 import { createStoredConversations } from './stored-searches';
 import { type AIMessage, type ToolCalls } from './types/AskiAi';
 import { EMPTY_TOOLS } from './utils/ai';
 
-import type { AgentStudioSearchParameters, AskAiSearchParameters, Memory, StoredAskAiState } from '.';
+import type { AgentStudioSearchParameters, Memory, StoredAskAiState } from '.';
 
 type UseChat = UseChatHelpers<AIMessage>;
 
@@ -27,21 +20,10 @@ type UseAskAiParams = {
   apiKey: string;
   appId: string;
   indexName: string;
-  useStagingEnv?: boolean;
-  searchParameters?: AskAiSearchParameters;
+  searchParameters?: AgentStudioSearchParameters;
   tools: ToolCalls;
-  agentStudio: boolean;
   memory?: Memory;
-} & (
-  | {
-      agentStudio: false;
-      searchParameters?: AskAiSearchParameters;
-    }
-  | {
-      agentStudio: true;
-      searchParameters?: AgentStudioSearchParameters;
-    }
-);
+};
 
 type UseAskAiReturn = {
   messages: AIMessage[];
@@ -81,51 +63,12 @@ const getAgentStudioTransport = ({
   });
 };
 
-const getAskAiTransport = ({
-  assistantId,
-  apiKey,
-  indexName,
-  searchParameters,
-  appId,
-  abortController,
-  useStagingEnv,
-}: Pick<UseAskAiParams, 'apiKey' | 'appId' | 'assistantId' | 'indexName' | 'searchParameters' | 'useStagingEnv'> & {
-  abortController: AbortController;
-}): DefaultChatTransport<AIMessage> => {
-  return new DefaultChatTransport({
-    api: useStagingEnv ? BETA_ASK_AI_API_URL : ASK_AI_API_URL,
-    headers: async (): Promise<Record<string, string>> => {
-      if (!assistantId) {
-        throw new Error('Ask AI assistant ID is required');
-      }
-
-      const token = await getValidToken({
-        assistantId,
-        abortSignal: abortController.signal,
-        useStagingEnv,
-      });
-
-      return {
-        ...(token ? { authorization: `TOKEN ${token}` } : {}),
-        'X-Algolia-API-Key': apiKey,
-        'X-Algolia-Application-Id': appId,
-        'X-Algolia-Index-Name': indexName,
-        'X-Algolia-Assistant-Id': assistantId || '',
-        'X-AI-SDK-Version': 'v5',
-      };
-    },
-    body: searchParameters ? { searchParameters } : {},
-  });
-};
-
 export const useAskAi: UseAskAi = ({
   assistantId,
   apiKey,
   appId,
   indexName,
-  useStagingEnv = false,
   tools = EMPTY_TOOLS,
-  agentStudio,
   searchParameters,
   memory,
 }) => {
@@ -133,24 +76,14 @@ export const useAskAi: UseAskAi = ({
 
   const askAiTransport = useMemo(
     () =>
-      agentStudio
-        ? getAgentStudioTransport({
-            apiKey,
-            appId,
-            assistantId: assistantId ?? '',
-            searchParameters,
-            userToken: memory?.userToken,
-          })
-        : getAskAiTransport({
-            assistantId: assistantId ?? '',
-            apiKey,
-            appId,
-            indexName,
-            searchParameters,
-            abortController: abortControllerRef.current,
-            useStagingEnv,
-          }),
-    [apiKey, appId, assistantId, indexName, useStagingEnv, agentStudio, searchParameters, memory?.userToken],
+      getAgentStudioTransport({
+        apiKey,
+        appId,
+        assistantId: assistantId ?? '',
+        searchParameters,
+        userToken: memory?.userToken,
+      }),
+    [apiKey, appId, assistantId, searchParameters, memory?.userToken],
   );
 
   // Sync ref during render so the stable `handleToolCall` (registered once
@@ -202,28 +135,19 @@ export const useAskAi: UseAskAi = ({
     async (messageId: string, thumbs: 0 | 1): Promise<void> => {
       if (!assistantId) return;
 
-      const res = await (agentStudio
-        ? postAgentStudioFeedback({
-            agentId: assistantId,
-            vote: thumbs,
-            messageId,
-            appId,
-            apiKey,
-            abortSignal: abortControllerRef.current.signal,
-          })
-        : postFeedback({
-            assistantId,
-            thumbs,
-            messageId,
-            appId,
-            abortSignal: abortControllerRef.current.signal,
-            useStagingEnv,
-          }));
+      const res = await postAgentStudioFeedback({
+        agentId: assistantId,
+        vote: thumbs,
+        messageId,
+        appId,
+        apiKey,
+        abortSignal: abortControllerRef.current.signal,
+      });
 
       if (res.status >= 300) throw new Error('Failed, try again later.');
       conversations.addFeedback?.(messageId, thumbs === 1 ? 'like' : 'dislike');
     },
-    [assistantId, agentStudio, appId, apiKey, useStagingEnv, conversations],
+    [assistantId, appId, apiKey, conversations],
   );
 
   const onStopStreaming = async (): Promise<void> => {
@@ -253,10 +177,8 @@ export const useAskAi: UseAskAi = ({
   const askAiError = useMemo((): Error | undefined => {
     if (!error) return undefined;
 
-    if (!agentStudio) return error;
-
     return getAgentStudioErrorMessage(error);
-  }, [error, agentStudio]);
+  }, [error]);
 
   return {
     messages,
